// frontend/src/backend.d.ts
export type Time = bigint;
// ...
export interface InterventionSlot {
    endHour: bigint;
    endMinute: bigint;
    startMinute: bigint;
    startHour: bigint;
}
// ...
export interface TimeEntry {
    id: string;
    heuresTrajet: bigint;
    date: Time;
    user: Principal;
    description: string;
    startMorning: bigint;
    endAstreinte?: bigint;
    interventionSlots: Array<InterventionSlot>;
    startAstreinte?: bigint;
    endMorning: bigint;
    endAfternoon: bigint;
    typeOfDay: DayType;
    heuresRepas: bigint;
    startAfternoon: bigint;
}
export enum DayType {
    conge = "conge",
    work = "work",
    astreinte = "astreinte"
}